requirements.txt
discord.py==2.3.2
Pillow==10.4.0
PyNaCl==1.5.0
